Move MakeReproPackage logic to a shared location
The move was pretty straightforward and I didn't really need to touch anything.

The only changes I made:
* crossgen2command.txt is now command.txt
* crossgen2repro.rsp is now repro.rsp
* Instead of "CrossGen2 version" the TXT file now says "Compiler version". What compiler will be obvious from the file name and version strings.
